+++ CHANGES.md
@@ -3,6 +3,7 @@
 - ci: Add missing dependency on Base64 and port to >=3.0.0 interface (#642 @avsm)
 - ci: Adapt to crunch.3.0.0 interface (#641 @hannesm)
 - ci: Use non-deprecated Yojson types in 1.7.0 and higher (#642 @avsm)
+- Support the WebMachine 0.6.0+ interface (@avsm)
 
 ### 0.12.2 (2019-01-18)